+++ CHANGELOG.md
@@ -1,7 +1,28 @@
 # Changelog
 
 
-## 1.19.3
+## 1.20.0
+
+### New
+
+* Add the ADD_CPES_IF_NONE option. [Ben Dalling]
+
+### Changes
+
+* Bump underlying Grype version from 0.48.0 to 0.49.0. [Ben Dalling]
+
+* Bump the version of Grype from 0.40.1 to 0.47.0. [Ben Dalling]
+
+### Fix
+
+* Ensure allowed bugs are registered as found. [Ben Dalling]
+
+### Other
+
+* Bump the version of Grype from 0.47.0 to 0.48.0. [Ben Dalling]
+
+
+## 1.19.3 (2022-06-28)
 
 ### New